some warning fixed and three logik errors
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@2335 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/components/gui_textinput.cc
@@ -61,7 +61,7 @@ void gui_textinput_t::infowin_event(const event_t *ev)
 					if (cursor_pos >= 0) {
 						cursor_pos = get_next_char(text, cursor_pos);
 					}
-					if (cursor_pos > strlen(text)) {
+					if (cursor_pos > (long)strlen(text)) {
 						cursor_pos = strlen(text);
 					}
 					break;
